Fix for #21
This fix is split into two parts.

1. Override undefined total_uri_count value to 0

The acquisition of the total_uri_count in the
clientContext so that a check for undefined can be performed.

In that case, I've overwritten the total_uri_count to be 0.

2.  Secondary check after schema validation

The schema validation should now pass as the total_uri_count is always
at least zero.

Just prior to submitting the ping, the total_uri_count is now checked
and if it is a zero - the ping is dropped.  This seems better than
sending up a bogus value.